requirements.txt
openai
python-dotenv
tiktoken
httpx
chromadb
sentence-transformers
pypdf
duckduckgo-search
pandas
numpy
matplotlib
streamlit
